email_verification_server/requirements.txt
aiofiles==0.8.0
aiosmtplib==1.1.7
anyio==3.6.1     
blinker==1.5     
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
gunicorn==20.1.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
twilio==7.12.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2